:art: 피드백을 반영하여 EmailViewController가 필요한 부분만 사용하게 했습니다.
